refactor(ssh): extract EnhancedSSHRunner to src/ssh/ssh_runner.py
- Move 2,745-line EnhancedSSHRunner class to dedicated module
- Remove FilePathUtils dependency (inline data dir logic)
- Fix self._get_data_directory() in static methods (was undefined)
- Add SSHConnectionConfig and SSHExecutionConfig dataclasses to module
- Remove unused imports from MistHelper.py (base64, getpass, hashlib, etc.)
- Exclude src/ssh/ from CI quality gates pending cleanup

Closes #201
