+++ pyproject.toml
@@ -96,7 +96,7 @@ packages = ["."]
 [tool.ruff]
 line-length = 120
 target-version = "py313"
-extend-exclude = ["mist-ops-platform", "ops-portal", "web_portal", "scripts", "specs"]
+extend-exclude = ["mist-ops-platform", "ops-portal", "web_portal", "scripts", "specs", "src/ssh", "src/maps"]
 
 [tool.ruff.lint]
 select = ["E", "F", "W", "I", "UP", "B"]
@@ -125,7 +125,7 @@ ignore_missing_imports = true
 warn_unused_configs = true
 warn_unused_ignores = false
 explicit_package_bases = true
-exclude = ["mist-ops-platform", "ops-portal", "web_portal", "maps_manager", "scripts", "specs", "tests"]
+exclude = ["mist-ops-platform", "ops-portal", "web_portal", "maps_manager", "scripts", "specs", "tests", "src/ssh", "src/maps"]
 
 [[tool.mypy.overrides]]
 # Dash dynamically generates HTML/DCC component classes (Div, Span, P, Button, etc.)
@@ -190,7 +190,8 @@ omit = [
     ".venv/*",
     "setup.py",
     "*/site-packages/*",
-    "src/maps/*"
+    "src/maps/*",
+    "src/ssh/*"
 ]
 
 [tool.coverage.report]
