+++ src/ssh/__init__.py
@@ -1,0 +1,1 @@
+"""SSH module for MistHelper remote command execution."""
